Add "3.7TB (44%) left of 8.4TB total" to success email
This slurps in a shim script called storage-check from the Datasafe 3 codebase.
Datasafe 3 runs on a dedicated NAS-like appliance and logs such lines to a persistent journal.
Datasafe 3 uses https://github.com/cyberitsolutions/journalcheck to turn that into an email.

datasafe-lite SHOULD NOT need this, because nagios should be checking ALL filesystems.
But some people run nagios check_disk with a whitelist of specific filesystems, so
do not notice when the datasafe-lite drive is nearly full.  Sigh.

Not checking inodes because "ran out of inodes!" is very rare.
Not checking before *AND* after because it would further obscure high-level flow in main().
Not reporting AT ALL in failure email, because
that would be fiddly (e.g. is it even mounted?) and it is just not worth it.